User

+++ src/main/java/com/moulliet/metro/DatasetResource.java
@@ -6,30 +6,21 @@ import com.moulliet.metro.load.GdbService;
 import com.sun.jersey.core.header.FormDataContentDisposition;
 import com.sun.jersey.multipart.FormDataParam;
 import org.apache.commons.io.IOUtils;
+import org.codehaus.jackson.map.ObjectMapper;
 import org.codehaus.jackson.node.ArrayNode;
+import org.codehaus.jackson.node.ObjectNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
+import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.*;
 
 @Path("/datasets")
 public class DatasetResource {
     private static final Logger logger = LoggerFactory.getLogger(DatasetResource.class);
-
+    private static final ObjectMapper mapper = new ObjectMapper();
 
     @GET
     @Produces("application/json")
@@ -41,8 +32,11 @@ public class DatasetResource {
     @PUT
     @Consumes("application/json")
     @Produces("application/json")
-    public Response putDatasets(String data) throws IOException {
-        logger.info("posted data {}", data);
+    public Response putDatasets(@HeaderParam("key") String key, String data) throws IOException {
+        logger.info("posted data {} key {}", data, key);
+        if (!Statics.userService.isValidKey(key)) {
+            return unauthorized();
+        }
         Statics.datasetService.update(data);
         Crashes.loadAll();
         return getDatasets();
@@ -52,6 +46,7 @@ public class DatasetResource {
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces("application/json")
     public Response uploadFile(
+            @HeaderParam("key") String key,
             @FormDataParam("datasetName") String datasetName,
             @FormDataParam("file") InputStream inputStream,
             @FormDataParam("file") FormDataContentDisposition contentDispositionHeader
@@ -59,7 +54,10 @@ public class DatasetResource {
             throws Exception {
         String fileName = contentDispositionHeader.getFileName();
         //todo - gfm - check for valid file name
-        logger.info("posting file {} as {}", fileName, datasetName);
+        logger.info("posting file {} as {} {}", fileName, datasetName, key);
+        if (!Statics.userService.isValidKey(key)) {
+            return unauthorized();
+        }
         File file = new File(Files.createTempDir() + fileName);
         OutputStream outpuStream = new FileOutputStream(file);
         IOUtils.copy(inputStream, outpuStream);
@@ -68,11 +66,21 @@ public class DatasetResource {
         return getDatasets();
     }
 
+    private Response unauthorized() {
+        ObjectNode node = mapper.createObjectNode();
+        node.put("message", "unauthorized access");
+        return Response.status(403).entity(node.toString()).build();
+    }
+
     @DELETE
     @Path("/{name}")
     @Produces("application/json")
-    public Response deleteDataset(@PathParam("name") String name) throws IOException {
+    public Response deleteDataset(@HeaderParam("key") String key,
+                                  @PathParam("name") String name) throws IOException {
         logger.info("deleting {}", name);
+        if (!Statics.userService.isValidKey(key)) {
+            return unauthorized();
+        }
         Statics.mongoDao.getDb().getCollection(name).drop();
         Statics.datasetService.delete(name);
         Crashes.loadAll();
